NOBUG: Refactor the JsonDataCleanup C# solution to a folder structure that will support multiple console apps
