+++ tools/legacy-scrape/scrape-tools/BCParks.ScrapeTools.JsonDataCleanup/BCParks.ScrapeTools.JsonDataCleanup.csproj
@@ -3,16 +3,18 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace>ProcessSeedData</RootNamespace>
+    <RootNamespace>BCParks.ScrapeTools.JsonDataCleanup</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="AngleSharp" Version="0.17.1" />
     <PackageReference Include="AutoMapper" Version="12.0.0" />
-    <PackageReference Include="HtmlAgilityPack" Version="1.11.46" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BCParks.ScrapeTools.Shared\BCParks.ScrapeTools.Shared.csproj" />
   </ItemGroup>
 
 </Project>
