+++ tools/legacy-scrape/scrape-tools/BCParks.ScrapeTools.JsonDataCleanup/Converters/ProtectedAreaCoordinatesConverter.cs
