+++ tools/legacy-scrape/scrape-tools/BCParks.ScrapeTools.JsonDataCleanup/Deserialization/ParkOperationSubareaDates.cs
