+++ tools/legacy-scrape/scrape-tools/BCParks.ScrapeTools.JsonDataCleanup/Deserialization/ProtectedAreaCoordinates.cs
