+++ tools/legacy-scrape/scrape-tools/BCParks.ScrapeTools.JsonDataCleanup/Serialization/ParkFacility.cs
