+++ tools/legacy-scrape/scrape-tools/BCParks.ScrapeTools.JsonDataCleanup/Serialization/ParkFireZoneXRef.cs
