+++ tools/legacy-scrape/scrape-tools/BCParks.ScrapeTools.JsonDataCleanup/Serialization/ParkOperationSubareaDates.cs
