+++ tools/legacy-scrape/visitor-services-patches/ScrapePatch -- ContentScrape line 402.vb
